INS-1644 playing around jets
