+++ core/jet.go
@@ -43,34 +43,3 @@ func (id JetID) Jet() (uint8, []byte) {
 	}
 	return id[PulseNumberSize], id[PulseNumberSize+1:]
 }
-
-// Parent returns a parent of the jet
-func (id JetID) Parent() JetID {
-	depth, prefix := id.Jet()
-	if depth == 0 {
-		return id
-	}
-
-	return *NewJetID(depth-1, ResetBits(prefix, depth-1))
-}
-
-// ResetBits returns a new byte slice with all bits in 'value' reset, starting from 'start' number of bit. If 'start'
-// is bigger than len(value), the original slice will be returned.
-func ResetBits(value []byte, start uint8) []byte {
-	if int(start) >= len(value)*8 {
-		return value
-	}
-
-	startByte := start / 8
-	startBit := start % 8
-
-	result := make([]byte, len(value))
-	copy(result, value[:startByte])
-
-	// Reset bits in starting byte.
-	mask := byte(0xFF)
-	mask <<= 8 - byte(startBit)
-	result[startByte] = value[startByte] & mask
-
-	return result
-}
