+++ ledger/storage/jet/tree.go
@@ -22,6 +22,7 @@ import (
 	"strings"
 
 	"github.com/insolar/insolar/core"
+	"github.com/insolar/insolar/ledger/storage"
 	"github.com/pkg/errors"
 	"github.com/ugorji/go/codec"
 )
@@ -248,7 +249,7 @@ func (t *Tree) Find(id core.RecordID) (*core.RecordID, bool) {
 	}
 	hash := id.Hash()
 	j, depth := t.Head.Find(hash, 0)
-	recID := core.RecordID(*core.NewJetID(uint8(depth), core.ResetBits(hash, depth)))
+	recID := core.RecordID(*core.NewJetID(uint8(depth), storage.ResetBits(hash, depth)))
 	return &recID, j.Actual
 }
 
@@ -277,8 +278,8 @@ func (t *Tree) Split(jetID core.RecordID) (*core.RecordID, *core.RecordID, error
 	}
 	j.Right = &jet{}
 	j.Left = &jet{}
-	leftPrefix := core.ResetBits(prefix, depth)
-	rightPrefix := core.ResetBits(prefix, depth)
+	leftPrefix := storage.ResetBits(prefix, depth)
+	rightPrefix := storage.ResetBits(prefix, depth)
 	setBit(rightPrefix, depth)
 	first := core.RecordID(*core.NewJetID(depth+1, leftPrefix))
 	second := core.RecordID(*core.NewJetID(depth+1, rightPrefix))
